yealink-module: Major rewrite to greatly improve performance (part 1)
The previous release suffered from a systematic bottleneck in the sysfs 
interface which would allow updates (eg. to the display) only in a 
fraction of the available time between key scans. The new version fixes 
this for the P1K.
A similar but less critical effect might have been observed on the P1KH, 
fixes will be provided in the next revisions (the current revision will 
not attach to a P1KH).

Further improvements have been made for the detection of the device 
model, which is now performed during probing. Any unknown devices will 
not be attached to the driver anymore!