yealink-module: Remove macros dev_dbg, dev_err and dev_warn.
These macros cause to suppress the real function of the kernel,
leading to a reduction in the details of the information listed.

The preprocessor test #ifndef is not applicable to verify the
availability of real function. And because all these functions
are available in all supported kernels >= 2.6.18, the tests and
macros are unnecessary and unwanted.

Signed-off-by: Slávek Banko <slavek.banko@axis.cz>